pyproject.toml
[tool.poetry]
name = "motorcycle-alert"
version = "0.1.0"
description = ""
authors = ["nbo2001 <natanael.oliveira@jusbrasil.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
pyTelegramBotAPI = "^4.14.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.1"
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"
pytest-cov = "^6.2.1"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
